Fix insane issue of main.tsx running twice and setting up ReactDOM root TWICE...?!?!
The reason I hunted down this issue was because every API request was appearing in the devtools twice. Somehow there was 2 instances of the app running, so every component render, every API request, every thing, was happening twice...

Not really a true fix but god knows where you'd start with that, maybe check the original template and what it's doing.
Doesn't appear to be an issue with serving prebuilt app.